Feat: Testando os endpoints, criado exception de validação de uso das entidades

+++ src/main/java/com/petshop/api/service/ProductCategoryService.java
@@ -4,10 +4,12 @@ import com.petshop.api.domain.validator.ValidatorEntities;
 import com.petshop.api.dto.request.CreateProductCategoryDto;
 import com.petshop.api.dto.request.UpdateProductCategoryDto;
 import com.petshop.api.dto.response.ProductCategoryResponseDto;
+import com.petshop.api.exception.EntitiesAlreadyInUseException;
 import com.petshop.api.exception.ResourceNotFoundException;
 import com.petshop.api.model.entities.ProductCategory;
 import com.petshop.api.model.mapper.ProductCategoryMapper;
 import com.petshop.api.repository.ProductCategoryRepository;
+import com.petshop.api.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -22,6 +24,8 @@ public class ProductCategoryService {
     private final ProductCategoryMapper productCategoryMapper;
     private final ProductCategoryRepository productCategoryRepository;
     private final ValidatorEntities validatorEntities;
+    private final ProductRepository productRepository;
+
 
     public ProductCategoryResponseDto getProductCategoryById(UUID id){
         return productCategoryRepository.findById(id)
@@ -58,6 +62,9 @@ public class ProductCategoryService {
         if (!productCategoryRepository.existsById(id)) {
             throw new ResourceNotFoundException("Category not found with ID: " + id);
             }
+        if (productRepository.existsByCategoryId(id)){
+            throw new EntitiesAlreadyInUseException("Cannot delete this category because it is being used by products");
+        }
         productCategoryRepository.deleteById(id);
     }
 }
